# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'LearnHub API Documentation'
  description: 'The documentation'
  version: 1.0.0
servers:
  -
    url: 'https://staging-api.learnhub.mk'
paths:
...
  '/content/blog-posts/{id}':
    get:
      summary: 'Blog post'
      operationId: blogPost
      description: 'This endpoint returns a single blog post'
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: ''
                  example: published
                  nullable: false
                  enum:
                    - draft
                    - published
                    - in_review
                    - archived
                publish_date:
                  type: string
                  description: 'Must be a valid date in the format <code>Y-m-d H:i:s</code>.'
                  example: '2024-12-01 02:38:27'
                  nullable: true
              required:
                - status
    put:
      summary: 'Update Blog post'
      operationId: updateBlogPost
      description: 'This endpoint will update a single blog post'
      parameters: []
      responses: {  }
      tags:
        - Content
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: ''
                  example: 'Example Blog Post Title '
                  nullable: false
                excerpt:
                  type: string
                  description: ''
                  example: 'This is test blogpost example'
                  nullable: false
                slug:
                  type: string
                  description: ''
                  example: 'example-blog-post-title '
                  nullable: false
                content:
                  type: string
                  description: ''
                  example: 'Lorem ipsum dolor sit amet, consectetur adipiscing ...'
                  nullable: false
                tags:
                  type: array
                  description: ''
                  example: '[1,2,3]'
                  items:
                    type: string
    delete:
      summary: 'Delete Blog posts'
      operationId: deleteBlogPosts
      description: 'This endpoint deletes blog post'
      parameters: []
      responses: {  }
      tags:
        - Content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: ''
                  example: archived
                  nullable: false
                  enum:
                    - draft
                    - published
                    - in_review
                    - archived
                publish_date:
                  type: string
                  description: 'Must be a valid date in the format <code>Y-m-d H:i:s</code>.'
                  example: '2024-12-01 02:38:27'
                  nullable: true
              required:
                - status
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the blog post.'
        example: non
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []